+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `Length::simplify` and automatically apply it to parsed expressions.
 * Change `presenter` and `presenter_opt` nodes to become the inner node if it is an widget.
 * Fix `force_size` layout causing an error log.
 * Fix `Scroll!` child alignment fill. Remove non standard layout, now only applies in dimensions without scrolling.